Extract Evaluation Metrics from Evaluation Job
